fix: delete user

+++ backend/authentication/views.py
@@ -261,11 +261,14 @@ class AccountDeleteView(generics.GenericAPIView):
     def delete(self, request, *args, **kwargs):
         id = kwargs.get('id')
         account = get_object_or_404(Account, id=id)
-        user = get_object_or_404(User, account=account)
-        user_data = model_to_dict(user)
-        user.delete()
-        if user_data.get('address'):
-            address = get_object_or_404(Address, id=user_data.get('address'))
-            address.delete()
+        user_data = None
+        if User.objects.filter(account=account).exists():
+            user = User.objects.get(account=account)
+            user_data = model_to_dict(user)
+            user.delete()
+        if user_data:
+            if user_data.get('address'):
+                address = get_object_or_404(Address, id=user_data.get('address'))
+                address.delete()
         account.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
